fix: correct ps and logs results after scale, on --status, and on odd ports (#1046)
Closes #1045

Four bugs on the read side of podup, plus a resilience follow-up caught
in review.

- **`logs` shows every replica after a `scale`.** It built its target
list from the static compose replica count, so `logs web` after `scale
web=3` streamed one of three. It now resolves against the live
containers (label-matched), falling back to the static names when a
service isn't up yet. Resolution is best-effort: if one selected
service's lookup errors, `logs` warns and streams the rest rather than
aborting — matching the per-container tolerance the command already
documents. Live replicas are sorted by their trailing index so
non-follow output stays in deterministic `-1, -2, -3` order.
- **`ps --status exited` returns the exited containers.** A status
filter now implies querying all containers, not just running ones —
libpod otherwise hides everything the filter is asking for, so the
command was always empty without `-a`.
- **`ps` port rendering can't overflow.** The host/container/range
arithmetic ran in `u16` on values that come straight from libpod's JSON;
a record like `host_port: 65535, range: 2` wrapped (garbage ports in
release, a panic under overflow-checks). It's widened now and renders
identically for normal values.
- **A restarting container no longer reports `Health: starting`.** The
health derivation matched the bare substring `starting`, which lives
inside `restarting`; it matches the real `health: starting` token only.

Test plan: introduced tests through the `fake_podman` Unix-socket
harness for each — every one was confirmed to fail against the pre-fix
code (the port-overflow tests panic under the repo's overflow-checked
dev profile). 1216 tests pass; clippy `-D warnings`, `fmt --check`, and
`cargo doc -D warnings` all clean. The podman-lane runs the whole thing
against real Podman 5 and 6.

One deliberate edge, consistent with the codebase's existing
404-tolerance: if EVERY selected service's live lookup errors, `logs`
exits 0 with no output and warnings on stderr, rather than a hard
failure — the same "show what exists" posture the per-container path
already takes.

---------
